refactor(macos): replace MTKView with CAMetalLayer + CAMetalDisplayLink
Fixes the focused-window rendering bug by moving off MTKView's
AppKit-coupled draw loop onto a plain NSView with a CAMetalLayer
backing, driven directly by CAMetalDisplayLink (macOS 14+).

Root cause of the bug: MTKView drives rendering via NSView.draw(),
which AppKit re-entered from dirty-rect / invalidation passes when
the window was focused and SwiftUI was actively invalidating layout.
That's why Mission Control's snapshot path rendered cleanly but the
focused live-composite path showed partial frames and artifacts.

Architectural changes:
- SpectrumRenderer: plain class owning all Metal state (device,
  queue, pipelines, textures, triple-buffer semaphore). No view
  coupling; unit-testable.
- MetalSpectrumNSView: NSView with CAMetalLayer as its backing
  layer. Owns the CAMetalDisplayLink, handles window attachment,
  backing-scale changes, and live resize.
- SpectrumWaterfallView: NSViewRepresentable now constructs a
  SpectrumRenderer + MetalSpectrumNSView pair; same public surface.
- SpectrumMTKView deleted.

Present-path specifics:
- Plain commandBuffer.present(drawable) — NOT present(_:atTime:).
  The atTime variant throws NSInvalidArgumentException from inside
  -[CAMetalDrawable presentWithOptions:] when used with a display-
  link-provided drawable on macOS 26; the display link's vsync
  cadence already aligns present timing.
- presentsWithTransaction = false (default). No more waitUntilScheduled
  main-thread block; the compositor-load symptom from the old code
  path is gone.
- Triple-buffer DispatchSemaphore(value: 3) gates in-flight frames,
  signalled from the command buffer's completion handler.

Verified: smooth 60 fps focused, unfocused, and in Mission Control
on M3 Pro (Mac15,7). Window title surfaces live FPS for quick
sanity-checking.
