+++ apps/macos/SDRMac/Renderer/SpectrumWaterfallView.swift
@@ -1,18 +1,18 @@
 //
-// SpectrumWaterfallView.swift — SwiftUI wrapper around
-// `SpectrumMTKView` via `NSViewRepresentable`.
+// SpectrumWaterfallView.swift — SwiftUI wrapper around the
+// CAMetalLayer-backed `MetalSpectrumNSView` via
+// `NSViewRepresentable`.
 //
-// SwiftUI doesn't host `NSView` subclasses natively;
-// `NSViewRepresentable` is the bridge. `makeNSView` runs once
-// per view lifetime; `updateNSView` runs on every binding
-// change and forwards the new values into the persistent
-// MTKView. The MTKView is NOT recreated on every binding update
-// — that would thrash Metal resources.
+// `makeNSView` runs once per view lifetime and wires up the
+// renderer. `updateNSView` runs on every binding change and
+// forwards new values into the persistent NSView. The NSView is
+// NOT recreated on every binding update — that would thrash
+// Metal resources.
 
 import SwiftUI
-// SdrCoreKit is not imported in sub-PR 1 — only the engine
-// wiring in sub-PR 3 needs it. Keeping the import out now
-// means a cleaner review of what's actually used.
+// SdrCoreKit is not imported yet — sub-PR 3 wires the real
+// engine. Keeping the import out for now means a cleaner review
+// of what's actually used.
 
 struct SpectrumWaterfallView: NSViewRepresentable {
     @Binding var minDb: Float
@@ -20,23 +20,20 @@ struct SpectrumWaterfallView: NSViewRepresentable {
 
     func makeNSView(context: Context) -> NSView {
         // Fail-soft: if Metal device creation / shader
-        // compilation fails, render a SwiftUI fallback
-        // message rather than crashing. Wrapping in an
-        // `NSHostingView` would be the pure-SwiftUI path; for
-        // simplicity we return a lightweight `NSView` with a
-        // red-ish label. Extremely rare failure mode — only
-        // happens on Macs without Metal support, which predate
-        // our macOS 14 floor anyway.
-        guard let mtk = SpectrumMTKView.make() else {
+        // compilation fails, render a SwiftUI-ish fallback
+        // rather than crashing. Extremely rare — only matters
+        // on Macs without Metal support, which predate our
+        // macOS 14 floor anyway.
+        guard let renderer = SpectrumRenderer.make() else {
             return makeFallbackView()
         }
-        mtk.applyBindings(minDb: minDb, maxDb: maxDb)
-        return mtk
+        let view = MetalSpectrumNSView(renderer: renderer)
+        view.applyBindings(minDb: minDb, maxDb: maxDb)
+        return view
     }
 
     func updateNSView(_ view: NSView, context: Context) {
-        guard let mtk = view as? SpectrumMTKView else { return }
-        mtk.applyBindings(minDb: minDb, maxDb: maxDb)
+        (view as? MetalSpectrumNSView)?.applyBindings(minDb: minDb, maxDb: maxDb)
     }
 
     private func makeFallbackView() -> NSView {
